added author info to package.json
