+++ .vscode/settings.json
@@ -1,0 +1,4 @@
+{
+  "comments.openView": "file",
+  "task.allowAutomaticTasks": "on"
+}
